refactor: api 파일 위치 변경 및 기존 api 파일 삭제
